move: 각 기능별 메인 페이지 분리
